% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_tax_and_benefits.R
\docType{data}
\name{IncomeTax}
\alias{IncomeTax}
\title{The Income Tax Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IncomeTax
}
\arguments{
\item{income}{The initial income of the person.}
}
\description{
This class describes the tax payable by an individual..
Income up to the personal allowance is tax free. Then income is taxed at
the rate in the rate vector until the corresponding threshold.
Note this class makes a number of simplifications including ignoring Scotland for sake of simplicity
}
\section{Fields}{

\describe{
\item{\code{personalAllowance}}{class used to calculate tax free personal alowance based upon income.}

\item{\code{threshold}}{The vector of tax thresholds, must be in ascending order.}

\item{\code{rate}}{The vector of tax rates, must be same size as thresholds.}
}}

\examples{
Income Tax
The Amount method

This method determines how much income tax a single person pays depending upon their income.city
amount()
}
\keyword{income}
\keyword{tax}
\keyword{uk}
