% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_tax_and_benefits.R
\docType{data}
\name{NationalInsurance}
\alias{NationalInsurance}
\title{The National Insurance class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NationalInsurance
}
\arguments{
\item{employeeThreshold}{A vector of thresholds for the employee}

\item{employeeRate}{The rates applied between the corresponding threshold and the next threshold}

\item{employerThreshold}{A vector of thresholds for the employee}

\item{employerRate}{The rates applied between the corresponding threshold and the next threshold}

\item{income}{The initial income of the person.}

\item{income}{The initial income of the person.}
}
\description{
The class captures the notion of UK national insurance
UK National insurance is a variable rate tax for both employees and employers
however because of the withdrawal of the personal allowance after 100000
UK income tax is not though it could be modeled as a variable rate tax.
}
\section{Fields}{

\describe{
\item{\code{employee}}{The variable rate tax structure for employee contributions.}

\item{\code{employer}}{The variable rate tax structure for employee contributions.}
}}

\examples{
Income Tax
The initialize method

Constructs UK national insurance
}
\keyword{Amount}
\keyword{Employee}
\keyword{Employer}
\keyword{Insurance}
\keyword{National}
\keyword{Security}
\keyword{Socail}
\keyword{The}
\keyword{This}
\keyword{UK}
\keyword{amount}
\keyword{an}
\keyword{be}
\keyword{by}
\keyword{determines}
\keyword{employee}
\keyword{employeeAmount()}
\keyword{employer}
\keyword{employerAmount()}
\keyword{given}
\keyword{income}
\keyword{income.}
\keyword{initial}
\keyword{initialize()}
\keyword{method}
\keyword{paid}
\keyword{tax}
\keyword{the}
\keyword{to}
\keyword{uk}
\keyword{with}
