% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_tax_and_benefits.R
\docType{data}
\name{PersonalAllowance}
\alias{PersonalAllowance}
\title{The Personal Allowance Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PersonalAllowance
}
\arguments{
\item{income}{The initial income of the person.}
}
\description{
This class describes the personal allowance that a person has in the UK tax system.
In the UK the personal allowance is how much you can earn tax free. However the allowance is 
withdrawn at the give rate over the threshold
Note this is a simplification of the actual UK tax system.
}
\section{Fields}{

\describe{
\item{\code{allowance}}{The amount of the allowance.}

\item{\code{threshold}}{The income at which the allowance starts to be withdrawn}

\item{\code{rate}}{The rate at which the allowance is withdrawn.}
}}

\examples{
Personal Allowance

This method determines how much personal alowance a person has dependant upon their income.ed
amount()
}
\keyword{allowance}
\keyword{income}
\keyword{personal}
\keyword{tax,}
\keyword{uk}
