% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_tax_and_benefits.R
\docType{data}
\name{TaxAndBenefits}
\alias{TaxAndBenefits}
\title{The Tax and Benefits  Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaxAndBenefits
}
\arguments{
\item{earnedIncome}{The initial earned income of the person.}

\item{unearnedIncome}{The initial unearned income of the person.
amount()
The Data Frame method

This method returns a data frame with the following columns based upon
the tax and benefit structures:
1) initialIncome
2) personalAllowance
3) incomeTax
4) employeesNationalInsurance
5) employersNationalInsurance
6) universalCreditEarned - assumes all income is earned
7) universalCreditUnearned - assumes all income is unearned
9) finalIncomeEarned - assumes all income is earned
10) finalIncomeUnearned - assumes all income is unearned}

\item{from}{The first point of initialIncome.}

\item{to}{The final point of the initial Income in the frame.}

\item{by}{The incriments by which initial income is increased in the frame.
dataFrame()}
}
\description{
This class describes the UK Tax and Benefits systeem
Note this is a simplification of the actual UK tax system.
}
\section{Fields}{

\describe{
\item{\code{incomeTax}}{The UK income tax structure}

\item{\code{nationalInsurance}}{The UK social security structure}

\item{\code{universalCredit}}{The UK asocial benifit structure}
}}

\examples{
Tax and Benefits
The amount method

This method determines the final amount a person will recieve after tax, national insurance
and universal credit.
}
\keyword{income}
\keyword{security}
\keyword{social}
\keyword{tax,}
\keyword{uk}
