% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_tax_and_benefits.R
\docType{data}
\name{UniversalCredit}
\alias{UniversalCredit}
\title{The Universal Credit Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UniversalCredit
}
\arguments{
\item{earnedIncome}{The initial earned income of the person.}

\item{unearnedIncome}{The initial unearned income of the person.}
}
\description{
This class describes the Universal Credit available to a single UK citizen
over 25 with no children.
}
\section{Fields}{

\describe{
\item{\code{amount}}{The amount of the credit.}

\item{\code{workAllowance}}{The amount that can be earned before credit is gradually withdrawn.}

\item{\code{earnedTaper}}{The rate at which credit is withdrawn for earnings above the workAllowance.}

\item{\code{unearnedTaper}}{The rate at which credit is withdrawn for unearned income.}
}}

\examples{
UniversalCredit
The amount method

This method determines how much universal credit is due dependant upon their income.
}
\keyword{allowance}
\keyword{amount()}
\keyword{credit}
\keyword{income}
\keyword{personal}
\keyword{tax,}
\keyword{uk}
\keyword{universal}
