% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_tax_and_benefits.R
\docType{data}
\name{VariableRateTax}
\alias{VariableRateTax}
\title{The Variable Rate Tax class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariableRateTax
}
\arguments{
\item{income}{The initial income of the person.}
}
\description{
A variable rate tax class captures the general idea of tax taken at different rates
for different levels of income.
UK National insurance is a variable rate tax for both employees and employers
however because of the withdrawal of the personal allowance after 100000
UK income tax is not though it could be modeled as a variable rate tax.
}
\section{Fields}{

\describe{
\item{\code{threshold}}{The vector of tax thresholds, must be in ascending order.}

\item{\code{rate}}{The vector of tax rates, must be same size as thresholds.}
}}

\examples{
Income Tax
The Amount method

This method determines the amount to be paid for the given income
given the thresholds and the rates.
}
\keyword{amount()}
\keyword{income}
\keyword{rate}
\keyword{tax}
\keyword{uk}
\keyword{variable}
