% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal-allowance.R
\docType{data}
\name{PersonalAllowance}
\alias{PersonalAllowance}
\title{The Personal Allowance Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PersonalAllowance
}
\description{
This class describes the personal allowance that a person has in the UK tax system.
In the UK the personal allowance is how much you can earn tax free. However the allowance is
withdrawn at the given rate over the threshold.
Note this is a simplification of the actual UK tax system. For example
this class ignores the married couples allowance.
}
\section{Fields}{

\describe{
\item{\code{allowance}}{The amount of the allowance.}

\item{\code{threshold}}{The income at which the allowance starts to be withdrawn.}

\item{\code{rate}}{The rate at which the allowance is withdrawn.}
}}

\section{Methods}{

\describe{
\item{\code{amount}}{Determines how much personal alowance a person has dependant upon their income.}
}
}

\examples{
personalAllowance2017 <- PersonalAllowance$new(
  base = 11500,
  threshold = 100000,
  withdrawalRate = 0.5
)
personalAllowance2017$amount(101000)
}
\keyword{allowance}
\keyword{income}
\keyword{personal}
\keyword{tax,}
\keyword{uk}
